Issue #1405: Improvements and fixes in accordance with comments
