+++ api/src/test/java/com/epam/pipeline/acl/issue/IssueApiServiceTest.java
@@ -16,10 +16,11 @@
 
 package com.epam.pipeline.acl.issue;
 
+import com.epam.pipeline.controller.PagedResult;
 import com.epam.pipeline.controller.vo.EntityVO;
 import com.epam.pipeline.controller.vo.IssueCommentVO;
 import com.epam.pipeline.controller.vo.IssueVO;
-import com.epam.pipeline.entity.datastorage.aws.S3bucketDataStorage;
+import com.epam.pipeline.entity.AbstractSecuredEntity;
 import com.epam.pipeline.entity.issue.Issue;
 import com.epam.pipeline.entity.issue.IssueComment;
 import com.epam.pipeline.entity.security.acl.AclClass;
@@ -30,6 +31,7 @@ import com.epam.pipeline.security.acl.AclPermission;
 import com.epam.pipeline.test.acl.AbstractAclTest;
 import com.epam.pipeline.test.creator.datastorage.DatastorageCreatorUtils;
 import com.epam.pipeline.test.creator.issue.IssueCreatorUtils;
+import com.epam.pipeline.test.creator.security.SecurityCreatorUtils;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.AccessDeniedException;
@@ -39,16 +41,19 @@ import java.util.Collections;
 import java.util.List;
 
 import static com.epam.pipeline.test.creator.CommonCreatorConstants.ID;
+import static com.epam.pipeline.test.creator.CommonCreatorConstants.TEST_INT;
 import static com.epam.pipeline.util.CustomAssertions.assertThrows;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.doReturn;
 
 public class IssueApiServiceTest extends AbstractAclTest {
 
-    private final EntityVO entityVO = IssueCreatorUtils.getEntityVO(ID, AclClass.DATA_STORAGE);
+    private static final AclClass ENTITY_ACL_CLASS = AclClass.DATA_STORAGE;
+
+    private final EntityVO entityVO = SecurityCreatorUtils.getEntityVO(ID, ENTITY_ACL_CLASS);
     private final Issue issue = IssueCreatorUtils.getIssue(entityVO, SIMPLE_USER);
     private final IssueVO issueVO = IssueCreatorUtils.getIssueVO(entityVO);
-    private final S3bucketDataStorage s3bucket = DatastorageCreatorUtils.getS3bucketDataStorage(ID, OWNER_USER);
+    private final AbstractSecuredEntity s3bucket = DatastorageCreatorUtils.getS3bucketDataStorage(ID, OWNER_USER);
     private final List<Issue> issueList = Collections.singletonList(issue);
     private final IssueComment issueComment = IssueCreatorUtils.getIssueComment(SIMPLE_USER);
     private final IssueCommentVO issueCommentVO = IssueCreatorUtils.getIssueCommentVO();
@@ -78,7 +83,7 @@ public class IssueApiServiceTest extends AbstractAclTest {
     public void shouldCreateIssueWhenPermissionIsGranted() {
         initAclEntity(s3bucket, AclPermission.READ);
         doReturn(issue).when(mockIssueManager).createIssue(issueVO);
-        doReturn(s3bucket).when(mockEntityManager).load(AclClass.DATA_STORAGE, ID);
+        doReturn(s3bucket).when(mockEntityManager).load(ENTITY_ACL_CLASS, ID);
         mockSecurityContext();
 
         assertThat(issueApiService.createIssue(issueVO)).isEqualTo(issue);
@@ -89,7 +94,7 @@ public class IssueApiServiceTest extends AbstractAclTest {
     public void shouldDenyCreateIssueWhenPermissionIsNotGranted() {
         initAclEntity(s3bucket);
         doReturn(issue).when(mockIssueManager).createIssue(issueVO);
-        doReturn(s3bucket).when(mockEntityManager).load(AclClass.DATA_STORAGE, ID);
+        doReturn(s3bucket).when(mockEntityManager).load(ENTITY_ACL_CLASS, ID);
         mockSecurityContext();
 
         assertThrows(AccessDeniedException.class, () -> issueApiService.createIssue(issueVO));
@@ -108,7 +113,7 @@ public class IssueApiServiceTest extends AbstractAclTest {
     public void shouldLoadIssueWhenPermissionIsGranted() {
         initAclEntity(s3bucket, AclPermission.READ);
         doReturn(issue).when(mockIssueManager).loadIssue(ID);
-        doReturn(s3bucket).when(mockEntityManager).load(AclClass.DATA_STORAGE, ID);
+        doReturn(s3bucket).when(mockEntityManager).load(ENTITY_ACL_CLASS, ID);
         mockSecurityContext();
 
         assertThat(issueApiService.loadIssue(ID)).isEqualTo(issue);
@@ -119,7 +124,7 @@ public class IssueApiServiceTest extends AbstractAclTest {
     public void shouldDenyLoadIssueWhenPermissionIsNotGranted() {
         initAclEntity(s3bucket);
         doReturn(issue).when(mockIssueManager).loadIssue(ID);
-        doReturn(s3bucket).when(mockEntityManager).load(AclClass.DATA_STORAGE, ID);
+        doReturn(s3bucket).when(mockEntityManager).load(ENTITY_ACL_CLASS, ID);
         mockSecurityContext();
 
         assertThrows(AccessDeniedException.class, () -> issueApiService.loadIssue(ID));
@@ -138,7 +143,7 @@ public class IssueApiServiceTest extends AbstractAclTest {
     public void shouldLoadIssuesForEntityWhenPermissionIsGranted() {
         initAclEntity(s3bucket, AclPermission.READ);
         doReturn(issueList).when(mockIssueManager).loadIssuesForEntity(entityVO);
-        doReturn(s3bucket).when(mockEntityManager).load(AclClass.DATA_STORAGE, ID);
+        doReturn(s3bucket).when(mockEntityManager).load(ENTITY_ACL_CLASS, ID);
         mockSecurityContext();
 
         assertThat(issueApiService.loadIssuesForEntity(entityVO)).isEqualTo(issueList);
@@ -149,7 +154,7 @@ public class IssueApiServiceTest extends AbstractAclTest {
     public void shouldDenyLoadIssuesForEntityWhenPermissionIsNotGranted() {
         initAclEntity(s3bucket);
         doReturn(issueList).when(mockIssueManager).loadIssuesForEntity(entityVO);
-        doReturn(s3bucket).when(mockEntityManager).load(AclClass.DATA_STORAGE, ID);
+        doReturn(s3bucket).when(mockEntityManager).load(ENTITY_ACL_CLASS, ID);
         mockSecurityContext();
 
         assertThrows(AccessDeniedException.class, () -> issueApiService.loadIssuesForEntity(entityVO));
@@ -223,7 +228,7 @@ public class IssueApiServiceTest extends AbstractAclTest {
         initAclEntity(s3bucket, AclPermission.READ);
         doReturn(issueComment).when(mockIssueManager).createComment(ID, issueCommentVO);
         doReturn(issue).when(mockIssueManager).loadIssue(ID);
-        doReturn(s3bucket).when(mockEntityManager).load(AclClass.DATA_STORAGE, ID);
+        doReturn(s3bucket).when(mockEntityManager).load(ENTITY_ACL_CLASS, ID);
         mockSecurityContext();
 
         assertThat(issueApiService.createComment(ID, issueCommentVO)).isEqualTo(issueComment);
@@ -235,7 +240,7 @@ public class IssueApiServiceTest extends AbstractAclTest {
         initAclEntity(s3bucket);
         doReturn(issueComment).when(mockIssueManager).createComment(ID, issueCommentVO);
         doReturn(issue).when(mockIssueManager).loadIssue(ID);
-        doReturn(s3bucket).when(mockEntityManager).load(AclClass.DATA_STORAGE, ID);
+        doReturn(s3bucket).when(mockEntityManager).load(ENTITY_ACL_CLASS, ID);
         mockSecurityContext();
 
         assertThrows(AccessDeniedException.class, () -> issueApiService.createComment(ID, issueCommentVO));
@@ -255,7 +260,7 @@ public class IssueApiServiceTest extends AbstractAclTest {
         initAclEntity(s3bucket, AclPermission.READ);
         doReturn(issueComment).when(mockIssueManager).loadComment(ID, ID);
         doReturn(issue).when(mockIssueManager).loadIssue(ID);
-        doReturn(s3bucket).when(mockEntityManager).load(AclClass.DATA_STORAGE, ID);
+        doReturn(s3bucket).when(mockEntityManager).load(ENTITY_ACL_CLASS, ID);
         mockSecurityContext();
 
         assertThat(issueApiService.loadComment(ID, ID)).isEqualTo(issueComment);
@@ -267,7 +272,7 @@ public class IssueApiServiceTest extends AbstractAclTest {
         initAclEntity(s3bucket);
         doReturn(issueComment).when(mockIssueManager).loadComment(ID, ID);
         doReturn(issue).when(mockIssueManager).loadIssue(ID);
-        doReturn(s3bucket).when(mockEntityManager).load(AclClass.DATA_STORAGE, ID);
+        doReturn(s3bucket).when(mockEntityManager).load(ENTITY_ACL_CLASS, ID);
         mockSecurityContext();
 
         assertThrows(AccessDeniedException.class, () -> issueApiService.loadComment(ID, ID));
@@ -325,5 +330,13 @@ public class IssueApiServiceTest extends AbstractAclTest {
         doReturn(issueComment).when(mockIssueManager).loadComment(ID, ID);
 
         assertThrows(AccessDeniedException.class, () -> issueApiService.deleteComment(ID, ID));
+    }
+
+    @Test
+    @WithMockUser
+    public void shouldLoadMyPagedResultListIssue() {
+        final PagedResult<List<Issue>> pagedResult = IssueCreatorUtils.getPagedListIssue();
+        doReturn(pagedResult).when(mockIssueManager).loadMy(ID, TEST_INT);
+        assertThat(issueApiService.loadMy(ID, TEST_INT)).isEqualTo(pagedResult);
     }
 }
